Add support for UnSet/Add/SetAll/UnSetAll
This completes the core features, including support for setting and
unsetting using an optional regex matching for the value to filter
the operation.

The behavior tries to preserve as much as possible the original
file contents, and automatically deletes empty sections, but preserves
them if they had user-created comments, for example.