Add support for e|edit action
Supports `core.editor` just like `git config`, and if not set,
will attempt to find VSCode in the path (using `which` or `where`
depending on the OS).

Fallback will just do a shell start of the target file, which
on Windows at least causes the app picker to show up.