Fix variable matching across multiple sections
We were improperly keeping the section around, which caused
multiple variables across multiple sections to match and failing
to set. Reworked a bit the variable matching, but this is
increasingly showing that the line-based approach is getting
too annoying and confusing to maintain.

We'll likely want to refactor the entire thing so that the
individual entries have a pointer to both the section and
the variable, so that we can properly backtrack from
variable to its containing section, and we should have a
separate section lookup for when we're adding variables.