Improve overload resolution for strings by using typed matcher
The overload receiving an optional string-valued regex was causing
the wrong overloads to be resolved when the `T` (in Get, GetAll and
others) was a string (i.e. from a default value) since all the other
arguments were also strings. So if we wanted to avoid passing in
the optional `subsection` value, we actually couldn't because the
extra string would be interpreted as the default value and not the
missing section (or the regex).

By making the regex a typed value, overload resolution for the
default value is more exact and less error-prone.

Also moving all usability overloads as extension methods to make
it more obvious they don't include any substantial code and are
just for API improvement.