Streamline and improve the main API entry point
The new `Config.Build` takes care of path-based building of
configuration as well as `ConfigLevel` enum for the known
System and Global locations. By having a single way, it's
easier to document and consume now. Also, we always provide
a local-path based file so that writing happens consistently
to that file, instead of (seemingly randomly) happen on the
local dir (if there was an existing file) or some parent
directory (if the file was found there instead).

Now there's 100% consistency in the behavior.