Fix minor issue with spaces when colons are used
While testing .netconfig with dotnet-sever using the
syntax for headers, I found out that for some reason,
the presence of the `:` char stops the parsing even if
what follows is a whitespace that should be allowed
by the existing parser. Adding the `.Or(Whitespace)`
somehow fixes it, which seems redundant but correct
nevertheless :/.

Added another test also that ensures no quotes are
needed for the mime-type mapping.