Fix packaging error from wrong output paths
Before we were multitargeting, we were setting the `AppendTargetFrameworkToOutputPath`
property to simplify the output dirs, but with multitargeting that now causes the
wrong files to be packaged in the nuget, since both TFMs now step on each other's
output directory.

This makes it so that the property is only set for the test project, where we don't
multitarget.