Fix issue when parsing a simple filename value
Since a value like `file.extension` is tokenized as a dotted identifier token instead of an identifier (or a string), we were not catching it in the variable value parser. Added another Or parser to catch that scenario too (and a previously failing test that now passes, to verify it).