Allow backslash without quotes by requiring escaping
This makes the behavior actually consistent with git-config, which also doesn't add quotes when the value contains a backslash, but it does require escaping the backslash regardless.

This means the only real use for quotes is when the value contains `#` or `;`, or when preserving before/after whitespaces is desired. This means that the vast majority of cases won't need them at all :)