Replace the parsing from Superpower to manual
Since the syntax is quite simple, it was simply overkill to use
Superpower in this case. It made understanding and even error
reporting more complicated, and the lack of a sufficiently complex
AST to represent the parsed lines made it even harder to justify.

This potentially simplifies integrating the library with others
who do not wish to take an extra (if indirect) dependency.

The new ConfigReader is a good foundation for potentially even
providing a source-only, read-only, forward-only, non-caching
version of the API for use strictly in runtime scenarios that
do not require the richer Config API.