Add support for .netconfig.user files
This allows to better separate local-only settings from potentially team-wide settings that can be commited to source control.

These files can be written to by using the new `ConfigLevel.Local` enum value for the mutating overloads receiving a level.

The overriding is possible at all the levels where the `.netconfig` is also valid, which provides the maximum flexibility.