Use the default file path instead of .user
Since we now prepend the `.user` local-only file to the aggregate list, we were using that as the FilePath in the aggregate config. It's more correct to use the non-user file since passing that when creating it, will still cause the .user to be prepended, so it's more of an override optional thing, rather than the main file.