When saving to an aggregate config, use fallback locations
We were previously assuming that there would always be a default config file (plain .netconfig in a non-system/global location) in the aggregate. This was problematic if you had opened a CLI app from the user's profile dir, which is even the default when running cmd, for example), since in that case there would be no default .netconfig, only the global config as the first entry.

So we instead use a smarter probing with fallbacks in place so that we try:

1. Default config if any found
2. Non-local (.user) file next
3. Local (.user) file if none found.

Fixes #51